requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
charset-normalizer==2.0.9
conda==4.3.16
distlib==0.3.4
Django==4.0
django-filter==21.1
django-versatileimagefield==2.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-flex-fields==0.9.6
filelock==3.4.0
idna==3.3
importlib-metadata==4.9.0
Markdown==3.3.6
mysqlclient==2.1.0
Pillow==8.4.0
pipenv==2021.11.23
platformdirs==2.4.0
pycosat==0.6.3
PyJWT==2.3.0
PyMySQL==1.0.2
python-magic==0.4.24
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
virtualenv==20.10.0
virtualenv-clone==0.5.7
zipp==3.6.0
